scraper/requirements.txt
scrapy==2.11.0
scrapy-playwright==0.0.40
playwright==1.43.0
psycopg2-binary==2.9.9
redis==5.0.4
flask==3.0.3
python-dotenv==1.0.1
phonenumbers==8.13.37
requests==2.31.0
fake-useragent==1.5.1
itemadapter==0.9.0
twisted==22.10.0
